Bring e2e/ and e2e-p2p/ under ESLint (type-aware)
Adds an e2e block to eslint.config.js modelled on wede's e2e block (the
fleet reference): same rules-of-hooks:off rationale (Playwright's use()
fixture arg collides with React 19's use-hook detection), same
no-unused-vars swap to the TS-aware rule. Unlike wede's block, and
unlike diwan's own untyped src block, this one is type-aware --
tseslint.configs.recommendedTypeChecked via parserOptions.projectService,
resolved against tsconfig.json (which already includes 'e2e' and
'e2e-p2p' from the earlier conversion commits). These specs are async
top-to-bottom, so no-floating-promises and friends catch a real flake
class (a dropped await) that the untyped rule set can't see.

Fixes the 9 real findings type-aware linting surfaced:
- stack.ts: attached a cause to a rethrown Error (preserve-caught-error,
  a core eslint:recommended rule in ESLint 10); dropped the pointless
  async on stop() (require-await) and retyped Stack.stop as () => void,
  updating both specs' now-synchronous stack.stop() call sites
  (no behavior change -- kill()/rmSync() were always synchronous, so the
  async wrapper awaited nothing).
- helpers.ts: removed a redundant "as string" (no-unnecessary-type-
  assertion) -- the value was already string once the earlier
  ": string | undefined" annotation was dropped (noUncheckedIndexedAccess
  is off repo-wide, so the indexed lookup already types as string).
- builtin-rendezvous-p2p.e2e.ts / rendezvous-p2p.e2e.ts: typed two
  .json() reads (no-unsafe-member-access) instead of letting any
  flow from Playwright's APIResponse.json().
- collab.e2e.ts: dropped an unused "const state =" capture
  (no-unused-vars) -- the return value was never read.
- import-open.e2e.ts: cast XLSX.write()'s return (typed any in the
  xlsx package's own .d.ts for every option combination) to the
  Uint8Array its docs promise for type: 'array', instead of letting
  that any propagate into makeXlsx/makeOds's Buffer return type
  (no-unsafe-return).
- slides-formatting.e2e.ts: one file-local disable for
  @typescript-eslint/unbound-method with the actual finding named -- the
  extracted native property setter is deliberately unbound so it can be
  invoked with an explicit receiver via .call(el, ...) on the next line.

+++ e2e/import-open.e2e.ts
@@ -52,18 +52,22 @@ async function makeDocx(bodyText: string): Promise<Buffer> {
   return zip.generateAsync({ type: 'nodebuffer' })
 }
 
+// xlsx's own types declare write()'s return as `any` (types/index.d.ts) for
+// every bookType/output combination — this asserts the concrete type its own
+// docs promise for `type: 'array'` (a Uint8Array), rather than letting that
+// `any` propagate into these functions' declared Buffer return type.
 function makeXlsx(cellText: string): Buffer {
   const wb = XLSX.utils.book_new()
   const ws = XLSX.utils.aoa_to_sheet([[cellText, 42], ['row2', 7]])
   XLSX.utils.book_append_sheet(wb, ws, 'Data')
-  return Buffer.from(XLSX.write(wb, { bookType: 'xlsx', type: 'array' }))
+  return Buffer.from(XLSX.write(wb, { bookType: 'xlsx', type: 'array' }) as Uint8Array)
 }
 
 function makeOds(cellText: string): Buffer {
   const wb = XLSX.utils.book_new()
   const ws = XLSX.utils.aoa_to_sheet([[cellText, 1], ['b', 2]])
   XLSX.utils.book_append_sheet(wb, ws, 'Data')
-  return Buffer.from(XLSX.write(wb, { bookType: 'ods', type: 'array' }))
+  return Buffer.from(XLSX.write(wb, { bookType: 'ods', type: 'array' }) as Uint8Array)
 }
 
 async function makePptx(text: string): Promise<Buffer> {
